Allow for node_modules path to be configurable
Prior to this commit we would assume the node_modules you were
attempting to rollup resided at process.cwd(), however if you are in a
deps of deps scenario (transitives) it may be required to use a different
resolution to node_modules. This makes this path configurable, by
allowing the caller to pass a fully qualified path to the node_modules
directory in which they want to resolve and rollup.